Zip file contains the whole code and story board for Student Library iOS application. Also attached a video where you can see the execution in iphone 11 pro max simulator.
In this iOS application, the user can add books that he borrowed from the library and can set a due date. Whenever due is being approached it will throw a notification to the user. Also, he can order/delete/view/rearrange the books that he added to the local database in the app.